adding a router
