+++ frontend/src/pages/DiscordTools.tsx
@@ -1,0 +1,18 @@
+export const DiscordTools = () => {
+  return (
+    <>
+      <header>
+        <div className="mx-auto max-w-7xl py-8 sm:px-6 lg:px-8">
+          <div className="min-w-0 flex-1 sm:px-6 lg:px-8">
+            <h1 className="text-3xl font-bold tracking-tight text-gray-900">
+              Discord Tools
+            </h1>
+          </div>
+        </div>
+      </header>
+      <main>
+        <div className="mx-auto max-w-7xl sm:px-6 lg:px-8"></div>
+      </main>
+    </>
+  );
+};
